Changed base salary from int to double

+++ src/PopsLaptopShop.java
@@ -19,7 +19,7 @@ public class PopsLaptopShop {
 
             //user information
             String name;
-            int baseSalary;
+            double baseSalary;
             String startingTier;
             int numBasic;
             int numPremium;
